+++ examples/preact-router/package.json
@@ -29,6 +29,9 @@
     "rimraf": "^3.0.2",
     "typescript": "^4.4.4",
     "vite": "^2.6.14",
-    "vite-plugin-pwa": "workspace:*"
+    "vite-plugin-pwa": "workspace:*",
+    "workbox-precaching": "^6.4.0",
+    "workbox-routing": "^6.4.0",
+    "workbox-window": "^6.4.0"
   }
 }
