+++ examples/preact-router/src/sw.ts
