+++ examples/preact-router/vite.config.ts
@@ -1,7 +1,55 @@
 import { defineConfig } from 'vite'
 import preact from '@preact/preset-vite'
-import { VitePWA } from 'vite-plugin-pwa'
+import { ManifestOptions, VitePWA, VitePWAOptions } from 'vite-plugin-pwa'
 import replace from '@rollup/plugin-replace'
+
+const pwaOptions: Partial<VitePWAOptions> = {
+  mode: 'development',
+  base: '/',
+  includeAssets: ['favicon.svg'],
+  manifest: {
+    name: 'PWA Router',
+    short_name: 'PWA Router',
+    theme_color: '#ffffff',
+    icons: [
+      {
+        src: 'pwa-192x192.png', // <== don't add slash, for testing
+        sizes: '192x192',
+        type: 'image/png',
+      },
+      {
+        src: '/pwa-512x512.png', // <== don't remove slash, for testing
+        sizes: '512x512',
+        type: 'image/png',
+      },
+      {
+        src: 'pwa-512x512.png', // <== don't add slash, for testing
+        sizes: '512x512',
+        type: 'image/png',
+        purpose: 'any maskable',
+      },
+    ],
+  },
+}
+
+const replaceOptions = { __DATE__: new Date().toISOString() }
+
+if (process.env.SW === 'true') {
+  pwaOptions.srcDir = 'src'
+  pwaOptions.filename = 'sw.ts'
+  pwaOptions.strategies = 'injectManifest'
+  ;(pwaOptions.manifest as Partial<ManifestOptions>).name = 'PWA Inject Manifest'
+  ;(pwaOptions.manifest as Partial<ManifestOptions>).short_name = 'PWA Inject'
+}
+else {
+  if (process.env.CLAIMS === 'true')
+    pwaOptions.registerType = 'autoUpdate'
+
+  if (process.env.RELOAD_SW === 'true') {
+    // @ts-ignore
+    replaceOptions.__RELOAD_SW__ = 'true'
+  }
+}
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -10,36 +58,8 @@ export default defineConfig({
     sourcemap: process.env.SOURCE_MAP === 'true',
   },
   plugins: [
-      preact(),
-    VitePWA({
-      mode: 'development',
-      base: '/',
-      registerType: process.env.CLAIMS === 'true' ? 'autoUpdate' : undefined,
-      includeAssets: ['favicon.svg'],
-      manifest: {
-        name: 'PWA Router',
-        short_name: 'PWA Router',
-        theme_color: '#ffffff',
-        icons: [
-          {
-            src: 'pwa-192x192.png', // <== don't add slash, for testing
-            sizes: '192x192',
-            type: 'image/png',
-          },
-          {
-            src: '/pwa-512x512.png', // <== don't remove slash, for testing
-            sizes: '512x512',
-            type: 'image/png',
-          },
-          {
-            src: 'pwa-512x512.png', // <== don't add slash, for testing
-            sizes: '512x512',
-            type: 'image/png',
-            purpose: 'any maskable',
-          },
-        ],
-      },
-    }),
+    preact(),
+    VitePWA(pwaOptions),
     replace({
       __DATE__: new Date().toISOString(),
       __RELOAD_SW__: process.env.RELOAD_SW === 'true' ? 'true' : 'false',
