+++ examples/react-router/package.json
@@ -34,8 +34,6 @@
     "typescript": "^4.4.4",
     "vite": "^2.6.14",
     "vite-plugin-pwa": "workspace:*",
-    "workbox-precaching": "^6.4.0",
-    "workbox-routing": "^6.4.0",
     "workbox-window": "^6.4.0"
   }
 }
