+++ examples/react-router/vite.config.ts
@@ -1,55 +1,7 @@
 import { UserConfig } from 'vite'
 import reactRefresh from '@vitejs/plugin-react-refresh'
-import { ManifestOptions, VitePWA, VitePWAOptions } from 'vite-plugin-pwa'
+import { VitePWA } from 'vite-plugin-pwa'
 import replace from '@rollup/plugin-replace'
-
-const pwaOptions: Partial<VitePWAOptions> = {
-  mode: 'development',
-  base: '/',
-  includeAssets: ['favicon.svg'],
-  manifest: {
-    name: 'PWA Router',
-    short_name: 'PWA Router',
-    theme_color: '#ffffff',
-    icons: [
-      {
-        src: 'pwa-192x192.png', // <== don't add slash, for testing
-        sizes: '192x192',
-        type: 'image/png',
-      },
-      {
-        src: '/pwa-512x512.png', // <== don't remove slash, for testing
-        sizes: '512x512',
-        type: 'image/png',
-      },
-      {
-        src: 'pwa-512x512.png', // <== don't add slash, for testing
-        sizes: '512x512',
-        type: 'image/png',
-        purpose: 'any maskable',
-      },
-    ],
-  },
-}
-
-const replaceOptions = { __DATE__: new Date().toISOString() }
-
-if (process.env.SW === 'true') {
-  pwaOptions.srcDir = 'src'
-  pwaOptions.filename = 'sw.ts'
-  pwaOptions.strategies = 'injectManifest'
-  ;(pwaOptions.manifest as Partial<ManifestOptions>).name = 'PWA Inject Manifest'
-  ;(pwaOptions.manifest as Partial<ManifestOptions>).short_name = 'PWA Inject'
-}
-else {
-  if (process.env.CLAIMS === 'true')
-    pwaOptions.registerType = 'autoUpdate'
-
-  if (process.env.RELOAD_SW === 'true') {
-    // @ts-ignore
-    replaceOptions.__RELOAD_SW__ = 'true'
-  }
-}
 
 const config: UserConfig = {
   // base: process.env.BASE_URL || 'https://github.com/',
@@ -58,7 +10,35 @@ const config: UserConfig = {
   },
   plugins: [
     reactRefresh(),
-    VitePWA(pwaOptions),
+    VitePWA({
+      mode: 'development',
+      base: '/',
+      registerType: process.env.CLAIMS === 'true' ? 'autoUpdate' : undefined,
+      includeAssets: ['favicon.svg'],
+      manifest: {
+        name: 'PWA Router',
+        short_name: 'PWA Router',
+        theme_color: '#ffffff',
+        icons: [
+          {
+            src: 'pwa-192x192.png', // <== don't add slash, for testing
+            sizes: '192x192',
+            type: 'image/png',
+          },
+          {
+            src: '/pwa-512x512.png', // <== don't remove slash, for testing
+            sizes: '512x512',
+            type: 'image/png',
+          },
+          {
+            src: 'pwa-512x512.png', // <== don't add slash, for testing
+            sizes: '512x512',
+            type: 'image/png',
+            purpose: 'any maskable',
+          },
+        ],
+      },
+    }),
     replace({
       __DATE__: new Date().toISOString(),
       __RELOAD_SW__: process.env.RELOAD_SW === 'true' ? 'true' : 'false',
